Get upstream patches (#4)
* GetNetworkInfo RPC edit

* Add issue template

* write_json_helper (un)pretty edit

* Json-rpc edit

Now returns 'false' if there is no 'params' member in the request.
Returns '"jsonrpc", "2.0"' by default.
There is still an issue with the 'id' member who is loosing his type w/ boost's ptree, so id is always returned as a string atm.

* Update boost-build.jam files

* Change gcc optimization to -O2 for libcoin #14

* Boost version check for property_tree paths

* Clean some warnings during build process

+++ coin/src/transaction.cpp
@@ -1011,8 +1011,8 @@ bool transaction::connect_inputs(
                     auto pindex = ptr_block_index;
                     pindex && ptr_block_index->height() - pindex->height() <
                     (constants::test_net ?
-                    constants::coinbase_maturity_test_network :
-                    constants::coinbase_maturity);
+                    static_cast<std::int32_t> (constants::coinbase_maturity_test_network) :
+                    static_cast<std::int32_t> (constants::coinbase_maturity));
                     pindex = pindex->block_index_previous()
                     )
                 {
